% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latParse}
\alias{latParse}
\title{Parse the latitude string in the B record}
\usage{
latParse(string)
}
\arguments{
\item{string}{The latitude string or vector of strings from the B record in DDMMmmmN/S format}
}
\value{
The latitude in decimal degrees.
}
\description{
Parse the latitude string in the B record
}
\examples{
latParse("1024361N")
}
