% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lonParse}
\alias{lonParse}
\title{Parse the latitude string in the B record}
\usage{
lonParse(string)
}
\arguments{
\item{string}{The longitude string or vector of strings from the B record in DDDMMmmmE/W format}
}
\value{
The longitude in decimal degrees.
}
\description{
Parse the latitude string in the B record
}
\examples{
lonParse("10243461E")
}
