% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse an igc file}
\usage{
parse(filepath, record = "B")
}
\arguments{
\item{filepath}{The filepath to the igc file.}

\item{record}{A string to select which record is to be parsed, defaults to "B".
Can be any record from A-L.}
}
\value{
The specified record from the igc file in a dataframe.
}
\description{
Parse an igc file
}
