% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recordLength}
\alias{recordLength}
\title{Give the amount of entries for a certain record}
\usage{
recordLength(charVec, record)
}
\arguments{
\item{charVec}{The vector of characters to operate on.}

\item{record}{The record to find the amount of entries for.}
}
\value{
The amount of entries for the given record.
}
\description{
Give the amount of entries for a certain record
}
\examples{
recordLength("test.igc","B")
}
